abstract products array into its own file and import where necessary, DRY out logic to determine which products to show based on inStockOnly value
